moved hexbinMap and ScatterPlot examples using glaPlot to separate files
